Rearranged code so that z_ functions are at the beginning of the import function.
The z_ functions have had a feature added so they will return all of the tokens they recognize -- this is used in the new cleanupBww2 function.

cleanupBww2:  Upgrades "old-style" ties to the new style by rewriting the token list.  This seemed like a better approach than modifying the rest of the parser.